requirements.txt
Package              Version
-------------------- --------
annotated-types      0.6.0
anyio                4.3.0
certifi              2024.2.2
click                8.1.7
colorama             0.4.6
dnspython            2.6.1
email_validator      2.1.1
fastapi              0.110.0
h11                  0.14.0
httpcore             1.0.4
httptools            0.6.1
httpx                0.27.0
idna                 3.6
itsdangerous         2.1.2
Jinja2               3.1.3
MarkupSafe           2.1.5
motor                3.3.2
orjson               3.9.15
pip                  23.2.1
pydantic             2.6.3
pydantic_core        2.16.3
pydantic-extra-types 2.5.0
pydantic-settings    2.2.1
PyJWT                2.8.0
pymongo              4.6.2
python-decouple      3.8
python-dotenv        1.0.1
python-multipart     0.0.9
PyYAML               6.0.1
sniffio              1.3.1
starlette            0.36.3
typing_extensions    4.10.0
ujson                5.9.0
uvicorn              0.27.1
watchfiles           0.21.0
websockets           12.0